Resolved version differences with MediatR and AutoMapper
I had an issue with MediatR and AutoMapper that was caused by version differences. This issue caused the registration of the two services to not happen and would trigger an error whenever the request was sent. I downgraded my versions to be compatible and my basic Database query went through