Create Request Working on server side
Create Video Game request working on server side with Unit Test built into it. After working on this feature I realized that I will have to jump back into these scripts as soon as I work on the validation feature.

Due to most of the of the data that is being sent through the requests are strings anything could pass through. It will be frustrating to create individual validations for each request., so I will build a feature around handling validations to expand coverage of my unit tests to ensure everything is working as intended.