Create View created (#12)
The Create Video Game functionality has been created on the React side. When I have a more solid internet connection I will be using copilot to help find mistakes and ensure that the dependances are correct for the tests in the client side. My testing done in the client side was unfortunately not in the github branch.

I also corrected the record error and made it so my command requests do not rely on a DTO class but instead are individualized on the Command itself.